Proof of concept: Remove debug statements like `Ember.assert`
This is a proof of concept to use the parsed syntax tree and remove all
debug statements like `Ember.[assert|deprecate|warn|debug]` or
`Ember.Logger.info`.

Currently those statements are required to be a oneliner and they are
removed by a simple string substitution (see http://git.io/i2wnpg).

By using something like this, it would be possible to even write
assertions like this:

```javascript
Ember.assert("this is a complex error", function() {
  if (state === "isError" && !stateHandled) { return false; }
  if (state === "warning" || state === "info") { return false; }

  // return true, to indicate that the assertion was OK
  return true;
});
```